requirements.txt
line-bot-sdk
gunicorn
flask
flask_httpauth
BeautifulSoup4
request
html5lib